Color Switcher
The changes adds a new button that sets the current color that the next class added to will be. The next class added should always be the color of that button, and clicking the button will rotate the specified color.

PS: I can't figure out how to run this website locally so I couldn't really test out the code. Is it possible to run this website locally so I can test it in future cases?